Convert BigInt<->Number based on library function `__sig` attributes
This change introduces and type to the `__sig` scheme that we using for
library functions: 'p'.  This letter signifies that the paramater or
return value is the width of a pointer which means it will differ
between wasm32 and wasm64.

For wasm64 we can then automatically convert these to int52/double at
the JS/Wasm boundary.