Remove unnecessary CatchInfo JS class. NFC
This class was used to hold the "adjusted" exceptions pointer along with
the original one.  If was separately allocated from the exception itself
but AFAICT this is now necessary.

Instead this adjusted pointer can simply be part of the __cpp_exception
class, as it is on other platforms. This not only simplifies the code
but also removes memory fragmentation and JS dependencies on the
allocate.